remove unused import to pass lint check
